crawler/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
openai==1.82.0
mysql-connector-python==8.3.0
python-dotenv==1.0.1
